Moving common testing functions/classes out to their own header
